requirements_sd15.txt
# U-REPA + SD-1.5 Requirements
# Python 3.9+

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.35.0

# Vision Models
timm>=0.9.0

# Data & Preprocessing
lmdb>=1.4.0
pillow>=10.0.0
pandas>=2.0.0
opencv-python>=4.8.0

# Training Infrastructure
accelerate>=0.25.0
peft>=0.7.0  # For LoRA
wandb>=0.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
einops>=0.7.0

# Evaluation
scikit-learn>=1.3.0  # For t-SNE, CCA
matplotlib>=3.7.0
seaborn>=0.13.0

# FID Calculation
clean-fid>=0.1.35  # Or pytorch-fid
scipy>=1.11.0

# CLIP (from OpenAI repo)
ftfy>=6.1.0
regex>=2023.0.0
git+https://github.com/openai/CLIP.git

# Optional: Image Quality Metrics
# lpips>=0.1.4  # For VAE reconstruction check
# torchmetrics>=1.2.0

# Development Tools (Optional)
# ipython>=8.12.0
# jupyter>=1.0.0
# black>=23.0.0
# flake8>=6.0.0
